requirements.txt
torch==2.2.0
torchvision==0.17.0

numpy==1.26.4
pandas==2.1.4

matplotlib==3.8.2
seaborn==0.13.2

scikit-learn==1.8.0

tqdm==4.66.2
Pillow==10.2.0